Updated the code relative to storing information and speech.
Created an array for storing answers.  Created functions for the robot's questions and the questions to test it. Set multiple timings. Created functions to let players speak and have the speech data recorded and reused differently. Adjusted the function to let the robot store answers and randomly pick an answer during knowledge checks.