Created the Ai jam program.
Begun work on the Ai jam program. Created arrays to store the possible items that could be requested and the items that are requested. Created a function to randomly select the items to request. Created the animation to display the list of items requested within the canvas. Added a basic object recognition function through ml5.