Modified level 3
Added extra chests to open, created a function to adjust the tracking of some cameras when a certain amount of chests are opened. Added a function to make the merchant and wizard dissapear when a certain amount of chests are open. Adjusted the attack function to only allow players to strike the rune if both the wizard and the merchant are gone